+++ forge/changelog.md
@@ -32,3 +32,4 @@
 - Improved native library loading and validation
 - Added `bypassGroupIsolation` to static audio channels
 - Prevent accidentally assigning left mouse button as push to talk in the voice chat settings
+- Added warning message when native libraries could not be loaded